Codebase audit: fix memory leak, add tracing, split large modules
- Fix pending request memory leak: wrap DashMap oneshot channels with
  PendingRequest<T> that tracks creation time, spawn periodic cleanup
  task removing stale entries older than 30s
- Add broadcast lag detection: new EventsLagged server event sent to
  GUI clients when broadcast channel overflows instead of silent drop
- Add request ID tracing middleware: UUID per request in x-request-id
  header and tracing span for log correlation (server + local mode)
- Split protocol/project.rs (1,369 LOC) into 7 focused submodules
  with full public re-exports for backward compatibility
- Split local/mod.rs (1,024 LOC) into router.rs and tasks.rs
- Add 5 integration tests for events_ws.rs (previously 0 tests)
- Add RFC documenting all audit findings
