+++ crates/zremote-agent/src/local/mod.rs
@@ -1,23 +1,19 @@
+mod router;
 mod routes;
 mod state;
+mod tasks;
 
 use std::net::SocketAddr;
 use std::path::PathBuf;
 use std::sync::Arc;
 use std::time::Duration;
 
-use axum::Router;
-use axum::routing::{delete, get, post};
-use tokio::sync::mpsc;
 use tokio_util::sync::CancellationToken;
-use tower_http::cors::CorsLayer;
-use tower_http::trace::TraceLayer;
 use uuid::Uuid;
-use zremote_protocol::{AgentMessage, AgenticAgentMessage};
 
-use crate::agentic::analyzer::{AnalyzerEvent, OutputAnalyzer};
-use crate::hooks::server::HooksServer;
+use router::build_router;
 use state::LocalAppState;
+use tasks::{spawn_agentic_detection_loop, spawn_pty_output_loop, start_hooks_server};
 
 /// Interval for periodic agentic tool detection (same as connection.rs).
 const AGENTIC_CHECK_INTERVAL: Duration = Duration::from_secs(1);
@@ -286,510 +282,6 @@ pub async fn run_local(
 
     tracing::info!("local mode stopped");
     Ok(())
-}
-
-fn build_router(state: Arc<LocalAppState>) -> Result<Router, Box<dyn std::error::Error>> {
-    let router = Router::new()
-        .route("/health", get(routes::health::health))
-        .route("/api/mode", get(routes::health::api_mode))
-        // Hosts endpoints (synthetic local host)
-        .route("/api/hosts", get(routes::hosts::list_hosts))
-        .route("/api/hosts/{host_id}", get(routes::hosts::get_host))
-        // Session CRUD
-        .route(
-            "/api/hosts/{host_id}/sessions",
-            post(routes::sessions::create_session).get(routes::sessions::list_sessions),
-        )
-        .route(
-            "/api/sessions/{session_id}",
-            get(routes::sessions::get_session)
-                .patch(routes::sessions::update_session)
-                .delete(routes::sessions::close_session),
-        )
-        .route(
-            "/api/sessions/{session_id}/purge",
-            delete(routes::sessions::purge_session),
-        )
-        .route(
-            "/api/sessions/{session_id}/context/push",
-            post(routes::sessions::push_context),
-        )
-        .route(
-            "/api/sessions/{session_id}/execution-nodes",
-            get(routes::sessions::list_execution_nodes),
-        )
-        .route(
-            "/api/execution-nodes/cleanup",
-            delete(routes::sessions::cleanup_execution_nodes),
-        )
-        .route(
-            "/api/sessions/previews",
-            get(routes::sessions::get_session_previews),
-        )
-        // Agentic loop endpoints
-        .route("/api/loops", get(routes::agentic::list_loops))
-        .route("/api/loops/{loop_id}", get(routes::agentic::get_loop))
-        // Projects
-        .route(
-            "/api/hosts/{host_id}/projects",
-            get(routes::projects::list_projects).post(routes::projects::add_project),
-        )
-        .route(
-            "/api/hosts/{host_id}/projects/scan",
-            post(routes::projects::trigger_scan),
-        )
-        .route(
-            "/api/hosts/{host_id}/browse",
-            get(routes::projects::browse_directory),
-        )
-        .route(
-            "/api/projects/{project_id}",
-            get(routes::projects::get_project)
-                .patch(routes::projects::update_project)
-                .delete(routes::projects::delete_project),
-        )
-        .route(
-            "/api/projects/{project_id}/sessions",
-            get(routes::projects::list_project_sessions),
-        )
-        .route(
-            "/api/projects/{project_id}/git/refresh",
-            post(routes::projects::trigger_git_refresh),
-        )
-        .route(
-            "/api/projects/{project_id}/worktrees",
-            get(routes::projects::list_worktrees).post(routes::projects::create_worktree),
-        )
-        .route(
-            "/api/projects/{project_id}/worktrees/{worktree_id}",
-            delete(routes::projects::delete_worktree),
-        )
-        .route(
-            "/api/projects/{project_id}/settings",
-            get(routes::projects::get_settings).put(routes::projects::save_settings),
-        )
-        // Project Actions
-        .route(
-            "/api/projects/{project_id}/actions",
-            get(routes::projects::list_actions),
-        )
-        .route(
-            "/api/projects/{project_id}/actions/{action_name}/run",
-            post(routes::projects::run_action),
-        )
-        .route(
-            "/api/projects/{project_id}/actions/{action_name}/resolve-inputs",
-            post(routes::projects::resolve_action_inputs_handler),
-        )
-        .route(
-            "/api/projects/{project_id}/prompts/{prompt_name}/resolve",
-            post(routes::projects::resolve_prompt),
-        )
-        .route(
-            "/api/projects/{project_id}/configure",
-            post(routes::projects::configure_with_claude),
-        )
-        // Config
-        .route(
-            "/api/config/{key}",
-            get(routes::config::get_global_config).put(routes::config::set_global_config),
-        )
-        .route(
-            "/api/hosts/{host_id}/config/{key}",
-            get(routes::config::get_host_config).put(routes::config::set_host_config),
-        )
-        // Knowledge
-        .route(
-            "/api/projects/{project_id}/knowledge/status",
-            get(routes::knowledge::get_status),
-        )
-        .route(
-            "/api/projects/{project_id}/knowledge/index",
-            post(routes::knowledge::trigger_index),
-        )
-        .route(
-            "/api/projects/{project_id}/knowledge/search",
-            post(routes::knowledge::search),
-        )
-        .route(
-            "/api/projects/{project_id}/knowledge/memories",
-            get(routes::knowledge::list_memories),
-        )
-        .route(
-            "/api/projects/{project_id}/knowledge/extract",
-            post(routes::knowledge::extract_memories),
-        )
-        .route(
-            "/api/projects/{project_id}/knowledge/generate-instructions",
-            post(routes::knowledge::generate_instructions),
-        )
-        .route(
-            "/api/projects/{project_id}/knowledge/write-claude-md",
-            post(routes::knowledge::write_claude_md),
-        )
-        .route(
-            "/api/projects/{project_id}/knowledge/bootstrap",
-            post(routes::knowledge::bootstrap_project),
-        )
-        .route(
-            "/api/projects/{project_id}/knowledge/generate-skills",
-            post(routes::knowledge::generate_skills),
-        )
-        .route(
-            "/api/projects/{project_id}/knowledge/memories/{memory_id}",
-            delete(routes::knowledge::delete_memory).put(routes::knowledge::update_memory),
-        )
-        .route(
-            "/api/hosts/{host_id}/knowledge/service",
-            post(routes::knowledge::control_service),
-        )
-        // Claude Tasks
-        .route(
-            "/api/claude-tasks",
-            post(routes::claude_sessions::create_claude_task)
-                .get(routes::claude_sessions::list_claude_tasks),
-        )
-        .route(
-            "/api/claude-tasks/{task_id}",
-            get(routes::claude_sessions::get_claude_task),
-        )
-        .route(
-            "/api/claude-tasks/{task_id}/resume",
-            post(routes::claude_sessions::resume_claude_task),
-        )
-        .route(
-            "/api/hosts/{host_id}/claude-tasks/discover",
-            get(routes::claude_sessions::discover_claude_sessions),
-        )
-        // Linear integration
-        .route(
-            "/api/projects/{project_id}/linear/me",
-            get(routes::linear::get_me),
-        )
-        .route(
-            "/api/projects/{project_id}/linear/issues",
-            get(routes::linear::list_issues),
-        )
-        .route(
-            "/api/projects/{project_id}/linear/issues/{issue_id}",
-            get(routes::linear::get_issue),
-        )
-        .route(
-            "/api/projects/{project_id}/linear/teams",
-            get(routes::linear::list_teams),
-        )
-        .route(
-            "/api/projects/{project_id}/linear/projects",
-            get(routes::linear::list_projects),
-        )
-        .route(
-            "/api/projects/{project_id}/linear/cycles",
-            get(routes::linear::list_cycles),
-        )
-        .route(
-            "/api/projects/{project_id}/linear/actions/{action_index}",
-            post(routes::linear::execute_action),
-        )
-        // Terminal WebSocket
-        .route(
-            "/ws/terminal/{session_id}",
-            get(routes::terminal::ws_handler),
-        )
-        // Events WebSocket
-        .route("/ws/events", get(routes::events::ws_handler))
-        .layer(TraceLayer::new_for_http())
-        .layer(CorsLayer::permissive())
-        .with_state(state);
-
-    Ok(router)
-}
-
-/// Start the hooks HTTP sidecar server.
-///
-/// This reuses the same `HooksServer` from the agent's server mode. Hook events
-/// from Claude Code arrive via HTTP, are translated to `AgenticAgentMessage`,
-/// and dispatched to the `AgenticProcessor` for local DB writes and event
-/// broadcasting.
-async fn start_hooks_server(state: Arc<LocalAppState>, shutdown: CancellationToken) {
-    // Channel for agentic messages from hooks
-    let (agentic_tx, agentic_rx) = mpsc::channel::<AgenticAgentMessage>(64);
-
-    // The hooks server needs an outbound_tx for AgentMessage (used for SessionIdCaptured).
-    // In local mode we don't need to send agent messages to a server, so we use a
-    // dummy channel that we drain and discard.
-    let (outbound_tx, _outbound_rx) = mpsc::channel::<AgentMessage>(64);
-
-    let sent_cc_session_ids = std::sync::Arc::new(tokio::sync::RwLock::new(
-        std::collections::HashSet::<String>::new(),
-    ));
-    let delivery_coordinator = std::sync::Arc::new(tokio::sync::Mutex::new(
-        crate::knowledge::context_delivery::DeliveryCoordinator::new(),
-    ));
-    let hooks_server = HooksServer::new(
-        agentic_tx,
-        state.session_mapper.clone(),
-        outbound_tx,
-        sent_cc_session_ids,
-        delivery_coordinator,
-    );
-
-    // Convert CancellationToken to a watch channel for the hooks server
-    let (shutdown_tx, shutdown_rx) = tokio::sync::watch::channel(false);
-    let shutdown_for_hooks = shutdown.clone();
-    tokio::spawn(async move {
-        shutdown_for_hooks.cancelled().await;
-        let _ = shutdown_tx.send(true);
-    });
-
-    match hooks_server.start(shutdown_rx).await {
-        Ok(addr) => {
-            tracing::info!(port = addr.port(), "hooks sidecar started");
-            // Install hooks into Claude Code settings
-            if let Err(e) = crate::hooks::installer::install_hooks().await {
-                tracing::warn!(error = %e, "failed to install CC hooks (non-fatal)");
-            }
-        }
-        Err(e) => {
-            tracing::warn!(error = %e, "failed to start hooks sidecar (non-fatal)");
-        }
-    }
-
-    // Spawn a task to consume agentic messages from hooks and dispatch to processor
-    spawn_hooks_message_consumer(state, agentic_rx);
-}
-
-/// Consume agentic messages from the hooks server channel and process them locally.
-fn spawn_hooks_message_consumer(
-    state: Arc<LocalAppState>,
-    mut agentic_rx: mpsc::Receiver<AgenticAgentMessage>,
-) {
-    tokio::spawn(async move {
-        while let Some(msg) = agentic_rx.recv().await {
-            // Register/unregister loop mappings
-            match &msg {
-                AgenticAgentMessage::LoopDetected {
-                    loop_id,
-                    session_id,
-                    ..
-                } => {
-                    state
-                        .session_mapper
-                        .register_loop(*session_id, *loop_id)
-                        .await;
-                }
-                AgenticAgentMessage::LoopEnded { loop_id, .. } => {
-                    tracing::info!(loop_id = %loop_id, "agentic loop ended (hook)");
-                    state.session_mapper.remove_loop(loop_id).await;
-                }
-                _ => {}
-            }
-
-            if let Err(e) = state.agentic_processor.handle_message(msg).await {
-                tracing::warn!(error = %e, "failed to process agentic hook message");
-            }
-        }
-    });
-}
-
-/// Spawn a periodic task that scans process trees for agentic tools.
-///
-/// Every 1 second, checks all active PTY sessions for known agentic tool
-/// processes (Claude Code, Codex, etc.) and emits detection/ended events.
-fn spawn_agentic_detection_loop(state: Arc<LocalAppState>) {
-    tokio::spawn(async move {
-        let mut interval = tokio::time::interval(AGENTIC_CHECK_INTERVAL);
-        // Skip the first immediate tick
-        interval.tick().await;
-
-        loop {
-            tokio::select! {
-                _ = interval.tick() => {
-                    // Collect session PIDs
-                    let session_pids: Vec<_> = {
-                        let mgr = state.session_manager.lock().await;
-                        mgr.session_pids().collect()
-                    };
-
-                    // Check for agentic tools
-                    let messages = {
-                        let mut mgr = state.agentic_manager.lock().await;
-                        mgr.check_sessions(session_pids.into_iter())
-                    };
-
-                    // Register/unregister loop mappings and dispatch to processor
-                    for msg in messages {
-                        match &msg {
-                            AgenticAgentMessage::LoopDetected {
-                                loop_id,
-                                session_id,
-                                ..
-                            } => {
-                                state.session_mapper.register_loop(*session_id, *loop_id).await;
-                            }
-                            AgenticAgentMessage::LoopEnded { loop_id, .. } => {
-                                tracing::info!(loop_id = %loop_id, "agentic loop ended (process exited)");
-                                state.session_mapper.remove_loop(loop_id).await;
-                            }
-                            _ => {}
-                        }
-
-                        if let Err(e) = state.agentic_processor.handle_message(msg).await {
-                            tracing::warn!(error = %e, "failed to process agentic detection message");
-                        }
-                    }
-
-                    // Check for idle loops and transition to WaitingForInput
-                    state.agentic_processor.check_idle_loops().await;
-                }
-                () = state.shutdown.cancelled() => {
-                    tracing::debug!("agentic detection loop shutting down");
-                    break;
-                }
-            }
-        }
-    });
-}
-
-/// Spawn a background task that reads PTY output from all sessions and routes it
-/// to the in-memory `SessionState` scrollback buffer and all connected browser
-/// WebSocket clients.
-///
-/// This is the local-mode equivalent of the PTY output handling in `connection.rs`.
-/// Instead of sending output over a WebSocket to a remote server, we write directly
-/// to the session store and browser senders.
-///
-/// Additionally, output is fed to the `AgenticLoopManager` for agentic state
-/// detection (e.g., Claude Code approval prompts, completion patterns).
-fn spawn_pty_output_loop(state: Arc<LocalAppState>) {
-    tokio::spawn(async move {
-        let mut pty_output_rx = state.pty_output_rx.lock().await;
-        let mut session_analyzers: std::collections::HashMap<
-            zremote_protocol::SessionId,
-            OutputAnalyzer,
-        > = std::collections::HashMap::new();
-
-        while let Some(pty_output) = pty_output_rx.recv().await {
-            let session_id = pty_output.session_id;
-            let data = pty_output.data;
-
-            if data.is_empty() {
-                // EOF from main pane -- session ended
-
-                // Clean up analyzer for this session
-                session_analyzers.remove(&session_id);
-
-                // Notify agentic manager that session closed
-                let loop_ended = {
-                    let mut mgr = state.agentic_manager.lock().await;
-                    mgr.on_session_closed(&session_id)
-                };
-                if let Some(msg) = loop_ended {
-                    if let AgenticAgentMessage::LoopEnded { ref loop_id, .. } = msg {
-                        state.session_mapper.remove_loop(loop_id).await;
-                    }
-                    if let Err(e) = state.agentic_processor.handle_message(msg).await {
-                        tracing::warn!(error = %e, "failed to process LoopEnded on session close");
-                    }
-                }
-
-                let exit_code = {
-                    let mut mgr = state.session_manager.lock().await;
-                    mgr.close(&session_id)
-                };
-                tracing::info!(
-                    session_id = %session_id,
-                    exit_code = ?exit_code,
-                    "PTY session ended"
-                );
-
-                // Update DB status
-                let session_id_str = session_id.to_string();
-                let _ = sqlx::query(
-                    "UPDATE sessions SET status = 'closed', exit_code = ?, closed_at = datetime('now') WHERE id = ?",
-                )
-                .bind(exit_code)
-                .bind(&session_id_str)
-                .execute(&state.db)
-                .await;
-
-                // Notify browser clients
-                {
-                    let mut sessions = state.sessions.write().await;
-                    if let Some(session_state) = sessions.get_mut(&session_id) {
-                        session_state.status = zremote_protocol::status::SessionStatus::Closed;
-                        let msg = zremote_core::state::BrowserMessage::SessionClosed { exit_code };
-                        session_state
-                            .browser_senders
-                            .retain(|tx| match tx.try_send(msg.clone()) {
-                                Ok(()) => true,
-                                Err(tokio::sync::mpsc::error::TrySendError::Full(_)) => true,
-                                Err(tokio::sync::mpsc::error::TrySendError::Closed(_)) => false,
-                            });
-                    }
-                }
-
-                // Remove from in-memory store
-                {
-                    let mut sessions = state.sessions.write().await;
-                    sessions.remove(&session_id);
-                }
-
-                // Broadcast event
-                let _ = state
-                    .events
-                    .send(zremote_core::state::ServerEvent::SessionClosed {
-                        session_id: session_id.to_string(),
-                        exit_code,
-                    });
-            } else {
-                // Feed through per-session analyzer
-                let analyzer = session_analyzers.entry(session_id).or_insert_with(|| {
-                    let default_cwd = dirs::home_dir().map(|p| p.to_string_lossy().to_string());
-                    OutputAnalyzer::with_initial_cwd(default_cwd)
-                });
-                let events = analyzer.process_output(&data);
-                for event in events {
-                    if let AnalyzerEvent::NodeCompleted(node) = event {
-                        let loop_id = {
-                            let mgr = state.agentic_manager.lock().await;
-                            mgr.loop_id_for_session(&session_id)
-                        };
-                        let msg = AgenticAgentMessage::ExecutionNode {
-                            session_id,
-                            loop_id,
-                            timestamp: node.timestamp,
-                            kind: node.kind,
-                            input: node.input,
-                            output_summary: node.output_summary,
-                            exit_code: node.exit_code,
-                            working_dir: node.working_dir,
-                            duration_ms: node.duration_ms,
-                        };
-                        if let Err(e) = state.agentic_processor.handle_message(msg).await {
-                            tracing::warn!(error = %e, "failed to process execution node");
-                        }
-                    }
-                }
-
-                // Main pane output -> scrollback + browser senders
-                let mut sessions = state.sessions.write().await;
-                if let Some(session_state) = sessions.get_mut(&session_id) {
-                    session_state.append_scrollback(data.clone());
-                    let msg = zremote_core::state::BrowserMessage::Output {
-                        pane_id: None,
-                        data,
-                    };
-                    session_state
-                        .browser_senders
-                        .retain(|tx| match tx.try_send(msg.clone()) {
-                            Ok(()) => true,
-                            Err(tokio::sync::mpsc::error::TrySendError::Full(_)) => true,
-                            Err(tokio::sync::mpsc::error::TrySendError::Closed(_)) => false,
-                        });
-                }
-            }
-        }
-    });
 }
 
 /// Insert or update the local host row in the database.
