+++ crates/zremote-cli/src/commands/events.rs
@@ -76,6 +76,7 @@ fn event_type_name(event: &ServerEvent) -> &'static str {
         ServerEvent::ExecutionNodeCreated { .. } => "execution_node_created",
         ServerEvent::ChannelPermissionRequested { .. } => "channel_permission_requested",
         ServerEvent::ChannelWorkerReply { .. } => "channel_worker_reply",
+        ServerEvent::EventsLagged { .. } => "events_lagged",
         ServerEvent::Unknown => "unknown",
     }
 }
