+++ crates/zremote-core/src/lib.rs
@@ -5,6 +5,8 @@ pub mod error;
 pub mod events_ws;
 pub mod processing;
 pub mod queries;
+#[cfg(feature = "axum")]
+pub mod request_id;
 pub mod snapshot_parser;
 pub mod state;
 #[cfg(feature = "axum")]
