+++ crates/zremote-protocol/src/project/mod.rs
@@ -1,0 +1,16 @@
+mod actions;
+mod git;
+mod info;
+mod linear;
+mod prompts;
+mod settings;
+
+#[cfg(test)]
+mod tests;
+
+pub use actions::*;
+pub use git::*;
+pub use info::*;
+pub use linear::*;
+pub use prompts::*;
+pub use settings::*;
